Updated the SF low level physics version of contact extensions with the normal filtering methods.
