--- Runtime/Scripts/Utilities/2D Utilities/ContactExtentions.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 9f98e7596c28ed1439a7fc169a7bb58b
