--- Runtime/Scripts/Utilities/2D Utilities/ContactFiltering.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: ae16c080074fc4943bd9cee05e53a583
